[Fixes #8499] Fix comments from review
