+++ geonode/services/tasks.py
@@ -54,7 +54,6 @@ def harvest_resource(self, harvest_job_id):
     try:
         handler = get_service_handler(
             base_url=harvest_job.service.service_url,
-            proxy_base=harvest_job.service.proxy_base,
             service_type=harvest_job.service.type
         )
         logger.debug("harvesting resource...")
