+++ geonode/services/utils.py
@@ -20,13 +20,8 @@
 import re
 import math
 import logging
-from typing import OrderedDict
 from django.conf import settings as django_settings
-from django.utils.translation import ugettext as _
 
-from geonode.services import enumerations
-from geonode.services.serviceprocessors.arcgis import ArcImageServiceHandler, ArcMapServiceHandler
-from geonode.services.serviceprocessors.wms import GeoNodeServiceHandler, WmsServiceHandler
 
 logger = logging.getLogger(__name__)
 
@@ -182,19 +177,3 @@ def parse_services_types():
             **custom_services_type_module.services_type
         }
     return custom_services_types
-
-
-def get_available_service_type():
-    default = OrderedDict({
-        enumerations.WMS: {"OWS": True, "handler": WmsServiceHandler, "label": _('Web Map Service')},
-        enumerations.GN_WMS: {"OWS": True, "handler": GeoNodeServiceHandler, "label": _('GeoNode (Web Map Service)')},
-        # enumerations.WFS: {"OWS": True, "handler": ServiceHandlerBase, "label": _('Paired WMS/WFS/WCS'},
-        # enumerations.TMS: {"OWS": False, "handler": ServiceHandlerBase, "label": _('Paired WMS/WFS/WCS'},
-        enumerations.REST_MAP: {"OWS": False, "handler": ArcMapServiceHandler, "label": _('ArcGIS REST MapServer')},
-        enumerations.REST_IMG: {"OWS": False, "handler": ArcImageServiceHandler, "label": _('ArcGIS REST ImageServer')},
-        # enumerations.CSW: {"OWS": False, "handler": ServiceHandlerBase, "label": _('Catalogue Service')},
-        # enumerations.OGP: {"OWS": True, "handler": ServiceHandlerBase, "label": _('OpenGeoPortal')},  # TODO: verify this
-        # enumerations.HGL: {"OWS": False, "handler": ServiceHandlerBase, "label": _('Harvard Geospatial Library')},  # TODO: verify this
-    })
-
-    return OrderedDict({**default, **parse_services_types()})
