+++ geonode/services/views.py
@@ -124,7 +124,7 @@ def _get_service_handler(request, service):
     feature many layers.
     """
     service_handler = get_service_handler(
-        service.service_url, service.proxy_base, service.type)
+        service.service_url, service.type)
     request.session[service.service_url] = service_handler
     logger.debug("Added handler to the session")
     return service_handler
